Add authorization to DeleteRestaurantCommandHandler
Introduced `ResourceOperation` enum to define CRUD operations.
Added `IRestaurantAuthorizationService` interface and its
implementation, `RestaurantAuthorizationService`, to handle
authorization logic for restaurant operations.

Updated `DeleteRestaurantCommandHandler` to include
authorization checks using `IRestaurantAuthorizationService`.
Throw `ForbidException` if the user is not authorized to delete
a restaurant.

Registered `IRestaurantAuthorizationService` in
`ServiceCollectionExtensions` for dependency injection.

Improved logging in `DeleteRestaurantCommandHandler` to log
successful restaurant deletions.

+++ Restaurants.Application/Features/Restaurants/Commands/DeleteRestaurantCommand.cs
@@ -1,17 +1,24 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Features.Restaurants.Commands;
 
 public record DeleteRestaurantCommand(int Id) : IRequest;
 
-public class DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository, ILogger<CreateRestaurantCommandHandler> logger) : IRequestHandler<DeleteRestaurantCommand>
+public class DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository, IRestaurantAuthorizationService restaurantAuthorizationService ,
+    ILogger<CreateRestaurantCommandHandler> logger) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
         var restaurant = await restaurantsRepository.GetByIdAsync(request.Id) ?? throw new NotFoundException("Restaurant", request.Id.ToString());
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbidException();
+
+
         await restaurantsRepository.DeleteAsync(restaurant);
         logger.LogInformation("Restaurant with Id {Id} deleted successfully.", request.Id);
